+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Fixed another duplication bug with a disconnected Crafting Grid (raoulvdberge)
 - Fixed oredict mode in autocrafting not working at all (raoulvdberge)
 - Removed getMissingItems() and getMissingFluids() functions from the OpenComputers integration, that info is now accessible through schedule(Fluid)Task(). If you just want to check if there are missing items/fluids but don't want to start an actual task, use the "canSchedule" parameter (raoulvdberge)
+- Removed the Interdimensional Upgrade, Network Transmitters are now cross dimensional by default (raoulvdberge)
+- Removed the per block FE cost of the Network Transmitter, it draws a fixed amount of FE/t now (raoulvdberge)
 - Updated Russian translation (kellixon)
 - Added fluid functions for the fluid autocrafting to the OpenComputers integration (raoulvdberge)